move tests, test times
