+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       - id: end-of-file-fixer
       - id: mixed-line-ending
       - id: name-tests-test
-        exclude: ^fastkml/tests/base.py
+        exclude: ^tests/base.py
       - id: no-commit-to-branch
       - id: pretty-format-json
       - id: requirements-txt-fixer
