+++ MANIFEST.in
@@ -7,3 +7,4 @@ include *.txt
 recursive-include docs *.py
 recursive-include docs *.rst
 recursive-include docs Makefile
+recursive-include tests *.py
