+++ tests/kml_test.py
@@ -16,8 +16,8 @@
 
 """Test the kml classes."""
 from fastkml import kml
-from fastkml.tests.base import Lxml
-from fastkml.tests.base import StdLibrary
+from tests.base import Lxml
+from tests.base import StdLibrary
 
 
 class TestStdLibrary(StdLibrary):
