+++ tests/styles_test.py
@@ -17,8 +17,8 @@
 """Test the styles classes."""
 
 from fastkml import styles
-from fastkml.tests.base import Lxml
-from fastkml.tests.base import StdLibrary
+from tests.base import Lxml
+from tests.base import StdLibrary
 
 
 class TestStdLibrary(StdLibrary):
