+++ tox.ini
@@ -32,6 +32,6 @@ max_line_length = 89
 ignore=
     W503
 per-file-ignores =
-    fastkml/tests/*.py: E722,E741,E501,DALL
+    tests/*.py: E722,E741,E501,DALL
     examples/*.py: DALL
 enable-extensions=G
